pyproject.toml
[project]
name = "docx-writer"
version = "0.1.0"
description = "Liteweight api for generating custom docx documents for SaMASZ"
authors = [
    {name = "SzymonIwaniuk",email = "s.f.iwaniuk@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
	"Django==5.2.5",
	"python-docx==1.2.0",
    "docx2pdf==0.1.8"
]

[tool.poetry]
packages = [{include = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
isort = "^6.0.1"
black = "^25.1.0"
pytest = "^8.3.5"
flake8 = "^7.0.0"
flake8-pyproject = "^1.2.3"

[tool.pytest.ini_options]
testpaths = [
    "tests/",
]

pythonpath = [
    "src",
]

[tool.flake8]
max-line-length = 120
extend-ignore = ["E203", "E401"]
exclude = [".venv"]
per-file-ingores = ["__init__.py"]

[tool.black]
line-length = 120
